+++ druid/src/widget/aspect_ratio_box.rs
@@ -139,13 +139,13 @@ impl<T: Data> Widget<T> for AspectRatioBox<T> {
         if bc.max() == bc.min() {
             warn!("Box constraints are tight. Aspect ratio box will not be able to preserve aspect ratio.");
 
-            return self.inner.layout(ctx, &bc, data, env);
+            return self.inner.layout(ctx, bc, data, env);
         }
 
         if bc.max().width == f64::INFINITY && bc.max().height == f64::INFINITY {
             warn!("Box constraints are INFINITE. Aspect ratio box won't be able to choose a size because the constraints given by the parent widget are INFINITE.");
 
-            return self.inner.layout(ctx, &bc, data, env);
+            return self.inner.layout(ctx, bc, data, env);
         }
 
         let bc = self.generate_constraints(bc);